feat: Share Module - squash commits
