+++ .github/workflows/build_linux.yml
@@ -46,7 +46,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - name: Install dependencies
-      run: sudo apt update && sudo apt-get install libgpac-dev
+      run: sudo apt update && sudo apt-get install libgpac-dev libnanomsg-dev
     - uses: actions/checkout@v4
     - name: run autogen
       run: ./autogen.sh
