+++ mac/Makefile.am
@@ -244,7 +244,7 @@ endif
 
 ccextractor_CFLAGS = -std=gnu99 -Wno-write-strings -Wno-pointer-sign -D_FILE_OFFSET_BITS=64 -DVERSION_FILE_PRESENT -DFT2_BUILD_LIBRARY -DGPAC_DISABLE_VTT -DGPAC_DISABLE_OD_DUMP -DGPAC_DISABLE_REMOTERY -DNO_GZIP
 
-ccextractor_LDFLAGS = $(shell pkg-config --libs gpac)
+ccextractor_LDFLAGS = $(shell pkg-config --libs gpac nanomsg)
 GPAC_CPPFLAGS = $(shell pkg-config --cflags gpac)
 
 ccextractor_CPPFLAGS =-I../src/lib_ccx/ -I../src/thirdparty/libpng/ -I../src/thirdparty/zlib/ -I../src/lib_ccx/zvbi/ -I../src/thirdparty/lib_hash/ -I../src/thirdparty/protobuf-c/ -I../src/thirdparty -I../src/ -I../src/thirdparty/freetype/include/
