+++ src/rust/lib_ccxr/src/lib.rs
@@ -2,9 +2,9 @@ pub mod activity;
 pub mod common;
 pub mod encoder;
 pub mod hardsubx;
+#[cfg(feature = "enable_sharing")]
+pub mod share;
 pub mod subtitle;
 pub mod teletext;
 pub mod time;
 pub mod util;
-#[cfg(feature = "enable_sharing")]
-pub mod share;
