+++ src/rust/lib_ccxr/src/share/tests.rs
@@ -1,13 +1,15 @@
 #![allow(unused_imports)]
 #![allow(unused)]
-use libc::c_char;
-use crate::share::share::*;
-use crate::share::ccxr_sub_entry_message::*;
-use std::sync::Once;
-use crate::util::log::{set_logger, CCExtractorLogger, DebugMessageFlag, DebugMessageMask, OutputTarget};
 
+#[cfg(feature = "enable_sharing")]
 mod test {
-    use super::*;
+    use crate::share::ccxr_sub_entry_message::*;
+    use crate::share::functions::sharing::*;
+    use crate::util::log::{
+        set_logger, CCExtractorLogger, DebugMessageFlag, DebugMessageMask, OutputTarget,
+    };
+    use std::os::raw::c_char;
+    use std::sync::Once;
 
     static INIT: Once = Once::new();
 
@@ -18,7 +20,7 @@ mod test {
                 DebugMessageMask::new(DebugMessageFlag::VERBOSE, DebugMessageFlag::VERBOSE),
                 false,
             ))
-                .ok();
+            .ok();
         });
     }
     #[test]
@@ -48,11 +50,7 @@ mod test {
             counter: 0,
             start_time: 0,
             end_time: 0,
-            lines: vec![
-                "test".to_string(),
-                "test".to_string(),
-                "test".to_string(),
-            ],
+            lines: vec!["test".to_string(), "test".to_string(), "test".to_string()],
         };
 
         unsafe {
@@ -117,7 +115,6 @@ mod test {
         }
     }
 
-
     #[test]
     fn test_ccxr_share_send() {
         initialize_logger();
@@ -138,7 +135,6 @@ mod test {
             time_out: 0,
             next: std::ptr::null_mut(),
             prev: std::ptr::null_mut(),
-
         };
         let status = unsafe { ccxr_share_send(&mut sub as *mut CcSubtitle) };
         assert!(matches!(status, CcxShareStatus::Ok | CcxShareStatus::Fail));
@@ -159,7 +155,6 @@ mod test {
         let status = unsafe { _ccxr_share_send(&*(&msg as *const CcxSubEntryMessage)) };
         assert!(matches!(status, CcxShareStatus::Ok | CcxShareStatus::Fail));
     }
-
 
     const CCX_DECODER_608_SCREEN_WIDTH: usize = 32;
     const CCX_DECODER_608_SCREEN_ROWS: usize = 15;
@@ -204,25 +199,50 @@ mod test {
             flags: 0,
             lang_index: 0,
             got_output: 1,
-            mode: [b'M' as c_char, b'O' as c_char, b'D' as c_char, b'E' as c_char, 0],
+            mode: [
+                b'M' as c_char,
+                b'O' as c_char,
+                b'D' as c_char,
+                b'E' as c_char,
+                0,
+            ],
             info: [b'I' as c_char, b'N' as c_char, b'F' as c_char, 0],
             time_out: 0,
             next: std::ptr::null_mut(),
             prev: std::ptr::null_mut(),
         };
-        let mut entries = CcxSubEntries { messages: Vec::new() };
+        let mut entries = CcxSubEntries {
+            messages: Vec::new(),
+        };
         let status = ccxr_share_sub_to_entries(&sub, &mut entries);
-        assert_eq!(status, CcxShareStatus::Ok, "Function should return OK status");
-        assert_eq!(entries.messages.len(), 1, "There should be one entry in messages");
+        assert_eq!(
+            status,
+            CcxShareStatus::Ok,
+            "Function should return OK status"
+        );
+        assert_eq!(
+            entries.messages.len(),
+            1,
+            "There should be one entry in messages"
+        );
 
         let message = &entries.messages[0];
         assert_eq!(message.start_time, 1000, "Start time should match input");
         assert_eq!(message.end_time, 2000, "End time should match input");
         assert_eq!(message.lines.len(), 3, "There should be 3 lines of content");
 
-        assert_eq!(message.lines[0], "Hello, World!", "First line content mismatch");
-        assert_eq!(message.lines[1], "Subtitle line 2", "Second line content mismatch");
-        assert_eq!(message.lines[2], "Subtitle line 3", "Third line content mismatch");
+        assert_eq!(
+            message.lines[0], "Hello, World!",
+            "First line content mismatch"
+        );
+        assert_eq!(
+            message.lines[1], "Subtitle line 2",
+            "Second line content mismatch"
+        );
+        assert_eq!(
+            message.lines[2], "Subtitle line 3",
+            "Third line content mismatch"
+        );
     }
 
     #[test]
@@ -243,18 +263,34 @@ mod test {
             flags: 0,
             lang_index: 0,
             got_output: 1,
-            mode: [b'M' as c_char, b'O' as c_char, b'D' as c_char, b'E' as c_char, 0],
+            mode: [
+                b'M' as c_char,
+                b'O' as c_char,
+                b'D' as c_char,
+                b'E' as c_char,
+                0,
+            ],
             info: [b'I' as c_char, b'N' as c_char, b'F' as c_char, 0],
             time_out: 0,
             next: std::ptr::null_mut(),
             prev: std::ptr::null_mut(),
         };
 
-        let mut entries = CcxSubEntries { messages: Vec::new() };
+        let mut entries = CcxSubEntries {
+            messages: Vec::new(),
+        };
 
         let status = ccxr_share_sub_to_entries(&sub, &mut entries);
 
-        assert_eq!(status, CcxShareStatus::Ok, "Function should return OK status");
-        assert_eq!(entries.messages.len(), 0, "There should be no messages for empty rows");
+        assert_eq!(
+            status,
+            CcxShareStatus::Ok,
+            "Function should return OK status"
+        );
+        assert_eq!(
+            entries.messages.len(),
+            0,
+            "There should be no messages for empty rows"
+        );
     }
 }
