+++ src/rust/src/libccxr_exports/mod.rs
@@ -1,10 +1,13 @@
 //! Provides C-FFI functions that are direct equivalent of functions available in C.
 
+
 pub mod bitstream;
 pub mod time;
 
+
 #[cfg(feature = "enable_sharing")]
 pub mod share;
+pub mod time;
 use crate::ccx_options;
 use lib_ccxr::util::log::*;
 use lib_ccxr::util::{bits::*, levenshtein::*};
